requirements.txt
python-telegram-bot==20.2
Random_Word==1.0.11
